Running Inference of FC 1 Hidden Layer NN on MNIST dataset
Ported a previous implementation of a Fully Connected 1 hidden layer
Neural Network with weights and bias created from a python script
using numpy and using SGD and backpropagation. Note that (1) this
inference was then subsequently ran on a beaglebone black with
success and (2) the python script is not uploaded. Next stage is
to completely port a batched SGD flow for training in here.